refact: move most key processing to new folder input/
This is a bit more controversial change; I'm getting fed up with everything being in a single keymap.h/cpp which is just getting more and more LOC. I wanted to split it to multiple files, but adding another file to src/ didn't feel like a good idea, as there was bound to be naming issues. I know that some people like having everything in single directory, but having background in Java, that feels just a bit too wrong for me :)

I ended up creating a new folder for key-processing related code, as most of it is relatively well encapsulated piece of logic anyway. Or maybe not 'encapsulated' due to all the globals, but a meaningful whole on its own, so to say.
